Bug fix and add JSON parser actions + JSON tree
In the parse_json_bool() function I accidentally had checked the
converted bool of 'null' in the union. This has been fixed.

Add several more actions to the parser. In particular:

    json_value:       json_object
		    | json_array
		    | JSON_STRING { $$ = *parse_json_string(ugly_text, &tree); }
		    | json_number
		    | JSON_TRUE { $$ = *parse_json_bool(ugly_text, &tree); }
		    | JSON_FALSE { $$ = *parse_json_bool(ugly_text, &tree); }
		    | JSON_NULL { $$ = *parse_json_null(ugly_text, &tree); }
		    ;

    json_number:    JSON_NUMBER { $$ = *parse_json_number(ugly_text, &tree); }
		    ;

In parse_json_null() check if converted is true. If not it's an error;
else print a debug statement saying it was converted.

In parse_json_number() it still does not print the number or check for
conversion. Part of this is there's more to check and I have to head off
for the day.

In parse_json_bool() there also was no change except for the bug fix as
above. However the function is now called when encountering JSON_TRUE or
JSON_FALSE.

The parse_json_member() function is not used yet and it's probable that
it'll have to change definition as well, taking a struct json *member.

JSON_OBJECT might be in a way mostly accounted for via the json_members
but not all is handled there and nothing is linked into the tree yet.
What this means is I believe (I might be wrong) is that except for
JSON_OBJECT and JSON_ARRAY conversion is done. What's not done is all
the checks on these and also linking it into the parse tree.

Added 'struct json tree' which is zero initialised. Right now the
parse_json_() functions are passed &tree but it's not used except to
check for NULL pointer. At a later commit at a later tree will be used.

There was a typo fix ('the the' by accident in json.c).

I hope that I documented everything here: there was a lot and I have to
stop working on this.