JSON parser updates
The json_member action now uses json_conv_member(). This means that the
parse_json_member() now is:

    struct json *parse_json_member(struct json *name, struct json *value, struct json *ast);

and the action looks like:

    +json_member:	JSON_STRING JSON_COLON json_element { $$ = *parse_json_member(&$1, &$3, &tree); }

which in the generated code is:

    { yyval = *parse_json_member(&yyvsp[-2], &yyvsp[0], &tree); }

Override default debug level in jparse to be DBG_MED and change
JSON_DBG_LEVEL to DBG_MED. I have the idea that a json debug function
could be added that is prefixed with JSON instead of dbg and which might
take a different variable which would mean that the other tools won't be
chatty but jparse can be. However for now this is not done. I made it
DBG_MED as I don't want too much information as there already is a lot
of information.

Updated some comments and remove duplicate XXX comments (some anyway).
There were comment fixes as well - the functions (except of the one for
arrays) do not return NULL pointers as of some commits ago.

TODO:

* The conversion routines should print better debug information. Right
now it only shows the type but it could also display the value - at
least for ones that have a value that's easy to define (strings being
the most obvious example).

* Add safer handling of union checking in struct json in the parse
functions. This will most likely be done in one of the next few commits
but it's possible it won't be today. But if it's not today it'll be in
the next few days.

* Write the parsing of arrays and verify that everything looks good.
After that we can start thinking about the parse tree and linking the
structs into it.

* Decide if the tree should be a global variable or not. I made it a
global variable yesterday to quickly get it in as I had to get going but
I was making use of the parse_json_() functions so I felt I should pass
in a tree.