Prefix json_element_type_name() names with JTYPE_
Also for the UNKNOWN one I made it: ((JTYPE_UNKNOWN)). I'm not sure if
the JTYPE_ should actually be there but I surrounded it with ()s to make
it clear(er) that it's not actually an enum type (though would only be
matched if a case was removed or a new enum type was added and not
handled yet).