Fix bug where json_member name was not parsed
In other words:

    $ ./jparse -s '{"foo":"bar"}' 2>&1|grep debug
    debug[0]: Calling parse_json_block("{"foo":"bar"}"):
    debug[0]: *** BEGIN PARSE:
    debug[3]: parse_json_string: about to parse string: <"foo">
    debug[3]: JSON return type: JTYPE_STRING
    debug[3]: parse_json_string: decoded string: <foo>
    debug[3]: parse_json_string: about to parse string: <"bar">
    debug[3]: JSON return type: JTYPE_STRING
    debug[3]: parse_json_string: decoded string: <bar>
    debug[3]: json_conv_member: JSON name type: JTYPE_STRING
    debug[3]: json_conv_member: JSON value type: JTYPE_STRING
    debug[3]: parse_json_member: convert member
    debug[0]: *** END PARSE

rather than only decoding the string "bar" as it previously did. The way
this was solved is by making JSON_STRING refer to json_string which
itself called parse_json_string() called rather than it being directly
in the JSON_STRING action in json_value.

Also changed the word 'decoded' to 'converted' in some of the other
functions.