Split JSON debug output from other debug output
New variable json_verbosity_level (in json_util) initialised to new
JSON_DBG_NONE which is defined (also in json_util):

    #define JSON_DBG_NONE   (DBG_NONE)      /* no JSON debugging information outside of the parser */
    #define JSON_DBG_LOW    (DBG_LOW)       /* minimal JSON debugging information outside of parser */
    #define JSON_DBG_MED    (DBG_MED)       /* somewhat more JSON debugging information outside of parser */
    #define JSON_DBG_LEVEL  (JSON_DBG_LOW)  /* default JSON debugging level json_verbosity_level */

Note that in jparse.y it is initially set to JSON_DBG_LEVEL which as can
be seen is JSON_DBG_LOW (which is actually DBG_LOW). This is done here
because later on we'll remove the initial override of the debugging
level.

An important point to be made about JSON_DBG_NONE: as the comment might
suggest the parser still generates debugging information and this
verbosity level is outside of that. Example location where json_dbg() is
used is the conversion functions in json.c. Once the parser is completed
we can disable (or limit) the parser verbosity.

New option to jparse: -J <level> which sets the json verbosity level.

With this commit there should be no need for other tools to worry about
additional debug information because the json_verbosity_level will be
set at 0. If any other tools will need this we can add the -J option to
support it. There is one problem with the choice of -J though: in
mkiocccentry both -j and -J are used (for jinfochk and jauthchk). For
the eventual jchk (which will perform both jinfochk and jauthchk
verifications) there might be a problem depending on the interface but
these both can be worked out later.