Clarify use of clearerr() / fclose() in lexer
Add a comment explaining why at the end of the function
parse_json_file() we directly use clearerr() or fclose() as the file
might be NULL (in fact it appears always to be so it's possible there's
a file descriptor leak: not sure, see below) and the
clearerr_or_fclose() makes that an error. Instead if stdin was read from
we use clearerr() directly on it; else if the file != NULL we fclose()
it (which seems to never happen).

Previously I noted in a commit that clearerr_or_fclose() might not be
needed but in fact it is and it's used in parse_json_file() just higher
up before the parsing.

As for a possible file descriptor leak: I'm not sure why calling
parse_json_block() which doesn't even act on ugly_in results in the file
being NULL (it instead acts on a YY_BUFFER_STATE) but it does for some
reason and that's why this is done this way (prior to using
parse_json_block() I used clearerr_or_fclose() on it at the end without
any problem). I'm sure there's some reason in the generated code but I'd
rather not have to go through it. Perhaps I could do a trace but this
seems unnecessary at least at this time so for now we do it this way.