Define UGLY__BUFFER_STATE to YY_BUFFER_STATE
Since the prefix change does not make a define for YY_BUFFER_STATE we do
it for the lexer. See the comment in jparse.l for more information. I
stress (once again) that this is a satire for the IOCCC though with no
apology; what we're sorry for is the appalling condition of the
(nonetheless useful) code that flex (and also bison) generates.