Allocate empty JSON arrays and objects when needed
In other words [] will allocate a JSON array properly and {} will
allocate a JSON object properly.

At the same time I've moved each to their respective rule. Previously it
looked like:

     json:		/* empty */
			| json_element
	-		| JSON_OPEN_BRACE JSON_CLOSE_BRACE
	-		| JSON_OPEN_BRACKET JSON_CLOSE_BRACKET
			;

but now the open and close brace is under json_object and open and close
bracket is under json_array.

The parsing of objects with members and arrays with elements is not
complete.