Make it clearer json_tree_walk() uses a callback
Rename func to callback to make it clearer that it's not just some
random and specific function but a function passed into the function.

Even though in C ALL functions are pointers which means one can do foo()
or (*foo)() we call callback() using the latter form to make it more
obvious that it is a function pointer (though the name callback might be
something of an indication as well).