Reorder JSON rules in jparse.y to match JSON spec
I have tried to make it as close to the specification order as possible
and I linked to the spec in the comments as well.

I'd also like to clarify commit 5923697d0e where I said:

        The json_member action now uses json_conv_member(). This means that the
	parse_json_member() now is:

	    struct json *parse_json_member(struct json *name, struct json *value, struct json *ast);

	and the action looks like:

	    +json_member:       JSON_STRING JSON_COLON json_element { $$ = *parse_json_member(&$1, &$3, &tree); }

	which in the generated code is:

	    { yyval = *parse_json_member(&yyvsp[-2], &yyvsp[0], &tree); }

Although it might be obvious what I meant it should have read that it
now uses parse_json_member() which itself uses json_conv_member() rather
than the other way round. Of course the '+' in front of json_member is
the diff and is not actually part of the grammar itself.

Also to clarify commit 626a984d where I wrote:

    ... rather than only decoding the string "bar" as it previously did.
    The way this was solved is by making JSON_STRING refer to
    json_string which itself called parse_json_string() called rather
    than it being directly in the JSON_STRING action in json_value.

This is what I call the '"change the thought in the middle of explaining
it but forget to update the rest of the thought" problem' which one can
compare with nonsense. What I was trying to say is that instead of
referring to the token JSON_STRING directly in the json_member rule we
refer to (at the time new) json_string which itself refers to the token
JSON_STRING. This solved the problem of only "bar" being parsed in the
json block:

    { "foo" : "bar" }