Fix double memory leak in assignment in txzchk
The code:

     file->basename = strdup(base_name(path)?base_name(path):"");

was problematic in a number of ways: base_name() actually returns a
duplicated copy of the passed in string (in this case path) with just
the basename. This means that I strdup()d a strdup()d string which means
the original strdup() reference is lost. Additionally because I used the
ternary operator to determine if it's != NULL only to call it again (if
!= NULL) means two references were lost and only the final strdup()
(file->basename) ends up being freed.