Fix compilation error under Fedora Linux
The following error was resolved:

    /usr/bin/ld: json_chk.o:json_chk.h:140: multiple definition of `found_common_json_fields'; mkiocccentry.o:json_chk.h:140: first defined here
    /usr/bin/ld: sanity.o:json_chk.h:140: multiple definition of `found_common_json_fields'; mkiocccentry.o:json_chk.h:140: first defined here
    collect2: error: ld returned 1 exit status
    make: *** [Makefile:345: mkiocccentry] Error 1

I am however aware of another problem under Fedora but it's only a
warning and unfortunately what I thought would solve it (see below) does
not and I unfortunately have to run the upgrade and shut the box down
after that as the UPS battery needs to be replaced and I don't have a
replacement at this time (as I have other things going on).

    cc: warning: json_chk.o: linker input file unused because linking not done

and

    cc: warning: json_util.o: linker input file unused because linking not done

Making the Makefile CFLAGS not include LDFLAGS and only adding LDFLAGS
to the targets that do not use -c does not solve the problem. At a look
at the man page it apparently does not have the option we use under
macOS (that might have different semantics)
-Wno-unused-command-line-argument and I'm not sure what else might
disable the warning and since I have to shut the box down it'll have to
be solved another time. At least now it can be compiled under Fedora.

Note that the tables found_(author|common|info)_json_fields might be
removed after the JSON parser is finished and the jinfochk and jauthchk
(and ultimately jchk) tools are finished but for now we need them in
because it allows make test to work (it might be that these tables are
kept in though just done differently - they have some value as they
stand and I have had some thoughts on how I might be able to make use of
them but this is down the road).