Add XXX todo note to jparse.{l,y}
Although I noted it in the issue on GitHub I think it's worth putting in
these files as well: once the parser is complete we should check as many
of the older versions of flex and bison as possible to determine the
oldest version that will work. For development I am under macOS which
(with MacPorts) has the most recent versions. However it's very possible
that older versions will work fine. I thought of testing it already and
then changing the version but since I work with macOS I don't want to
accidentally introduce a problem before it's even finished.

Also to clarify something with commit 54587657 where I said how the
tables found_(author|common|info)_json_fields might be removed after the
parser is complete. This paragraph should have been above the discussion
of the compilation warning (and see below) under Fedora; the warning is
not related to the tables but rather the fact that not all files need
-lm which causes the warning indicated in the cited commit.

Note that doing:

    make CC=clang all

will not trigger this warning as clang has the option to disable that
warning and gcc does not appear to have it (at least it does not under
Fedora: CentOS does not seem to have the warning at all or if it does
it's less picky about it). The warning should be addressed but it's a
harmless warning and can safely be ignored. As I said I have to keep the
fedora box down most the time so it's harder to fix this but I'll play
with it some later on at some point.