Fix linker warning under Fedora Linux
Whilst the system downloads the upgraded packages I looked at this a bit
more. The warning is clearer than my head was at the time: it wasn't a
matter of linker flags but rather linked in files. Specifically sanity.o
does not need to link in json_chk.o and jstrencode.o does not need to
link in json_util.o.

Thus the warning described in commit 54587657 has been resolved.