Typo fixes and minor changes to ugly_ satire
In the comments for the (not ugly but beautiful) satire for the IOCCC
(the ugly_ prefix) there were some minor fixes and changes.

In one place in both files (due to yank paste) I had typed 'addition'
instead of 'addiction'.

I also changed some stress on some words (adding and changing where the
stress is ... though if we're looking for actual stress looking at and
trying to work with the generated code would be much more stressful).