bool <-> string macro updates
The function string_to_bool() is now a macro called strtobool() (the
name is analogous to the strtol() etc.). It looks like:

    #define strtobool(x) ((x) != NULL && !strcmp((x), "true"))

and is in both dbg.h and util.h (the latter of which now #includes
string.h) which means that if x != NULL and is exactly the string "true"
it will return true (or more correctly 1) and otherwise it will be false
(0).

The t_or_f() macro has been renamed booltostr() (the analogous inverse
of strtol() etc.) and has not been changed in definition.

These names were carefully selected to make the most sense. Perhaps
string_to_bool() and bool_to_string() would make more sense but the
former in part brought up this discussion in the first place so I went
with what (most) matches the C library.