Extra sanity checks added to parse_json_bool
Make sure that the allocated string != NULL (should this be done for the
other functions?).

Make sure that the allocated string is either "true" or "false".

Convert the converted boolean back to a string and then:

- Verify that it's != NULL
- That the string is equal (!strcmp() to the allocated string.
- That the allocated string (when converted to a boolean) is equal to
  the converted value.
- And make sure that the converted value (as a string i.e. from
  booltostr()) as a boolean is equal to the converted value of the
  original string.

Only if these tests pass is the conversion considered valid. Maybe this
is overkill but this makes absolute certain that there were no faults in
the conversion (or almost certain - I can actually think of another
couple checks that could be added).