Reformat JSON parser rules and actions a bit
Move the | for OR to the end of the respective lines (rather than
in front of the next).

Make actions look more like:

    json_string:	JSON_STRING
			{
			    $$ = *parse_json_string(ugly_text, &tree);
			}
			;

for easier reading when it's more than one line of actions (though for
now there are none).

When there are three parts of a rule e.g. elements break up the rule so
it's clearer that there are parts e.g. for json_elements:

    json_elements:	json_element	|
			json_element
			    JSON_COMMA
			json_elements
			;

Note that these are subject to change.