Added lots of json_dbg() calls to bison grammar
In an effort to better understand the flow the bison JSON parser,
we added (temporally?) calls to json_dbg() at every stage throughout
the bison grammar.

Now you can do commands such as:

    ./jparse -J1 -s '[ 1, { "foo": false }, [ 3 ], null ]' 2>&1 | fgrep 'JSON DEBUG['

Sorry (tm Canada🇨🇦) to botch / mess up for formatting of the bison
grammar.  It was hard to determine the best way to inject all those
calls to json_dbg().