Change to make code indentation consistent
While we are not sure of the best way to format bison grammar,
we are sure that whatever the style, it should be consistent.
This is a minor change to make the indention consistent.

At least if the formatting needs to change later on,
it can be changed consistently thought the grammar.  :-)