Fix JSON_DBG_LEVEL in jparse.{l,y}
These are now json_verbosity_level which happens to still be
JSON_DBG_LEVEL in jparse. The way this came about is as follows:

Prior to having the json_dbg() facility it was desired to not have debug
level 0 as this would affect other json tools as well. I thus added
JSON_DBG_LEVEL and made the calls use that so as to not clutter up the
other tools (I believe I had set the verbosity_level to JSON_DBG_LEVEL
in jparse so that I actually would see it). But then I added the json_dbg()
facility and also made json_verbosity_level which I still kept at
JSON_DBG_LEVEL (it can be overridden by -J option in jparse). The
function calls were changed to use the new facility but JSON_DBG_LEVEL
was not changed to json_verbosity_level.