Fix JSON debug output in JSON conversion functions
There might have been a misunderstanding in how json_dbg() works wrt the
json_verbosity_level: it works just like the regular verbosity_level and
dbg() which means unless it's set to a higher value (like done directly
in jparse or the -J option in jparse) using json_verbosity_level in
calls to json_dbg() will be 0 which means there will be output that
might not be needed.

The -J json_level option has been added to jnum_chk and also
ioccc-test.sh (default 0 for both). Thus if one does:

    ./ioccc-test.sh -J 1

they will see lines like:

    JSON DEBUG[1]: JSON return type: JTYPE_NUMBER

but if they do not (or they specify -J 0) they will not see that output.
The JSON_DBG_LOW in the json conversion functions was arbitrarily
selected and could be changed.