Fixes in jparse.y
The functions:

    struct json *parse_json_string(char const *string);
    struct json *parse_json_number(char const *string);
    struct json *parse_json_bool(char const *string);
    struct json *parse_json_null(char const *string);

no longer take a struct json *ast. The other two might not need it as
well but for now they still take it.

When conversion fails in any of the parse_json_() functions it is now an
error rather than a warning.

Comments have been updated to reflect the above changes but it's very
possible that I've missed some things as I did not look at the comments
in detail.