Removed setting of ugly_lval->type in flex scanner
From the context that we can see in the debugger, ugly_lval
was a NULL pointer and this was invalid within the scanner's
context.  So we removed assigning ugly_lval->type in scanner
and will leave setting of the struct json type up to
the parser and its helper friend functions.