jparse updates (comments and WHITESPACE -> WS)
The comments reflect more recent developments with the parser including
no longer stating that there are no actions (in both flex and bison
files).

Although I prefer JSON_WHITESPACE the spec uses ws so we use JSON_WS
(eventually we won't even need to do anything with it so this won't be
an issue at all).