Resolve case of vim spellcap in json.c
To satisfy the grammar check I've surrounded a . with 's so that vim
doesn't think it's a new sentence that does not start with a capital
letter (for the logophiles who do not know :) this is also known as an
uncial or a majuscule). As a general rule with text like:

	    "%s: e notation numbers cannot have . before e or E: <%s>"

the dot should be surrounded by 's to help prevent this problem. It also
makes it clearer, I think, what exactly the problem is (this
unfortunately cannot be solved automatically as there are many places
where this should not be done).