Add initial json_parse() function
This function is incomplete! All it does currently is some sanity checks
on the pointer and length. It still has to make use of parse_json_block
(which maybe should be renamed parse_json_blob) which itself might have
to change definition a bit to pass length in (it's also possible that
this function will replace that function and the parse_json_file() will
call this one instead but this is yet to be determined).

The function looks like:

    /* json_parse
     *
     * Given a pointer to char and a length, use the parser to determine if the json
     * is valid or not.
     *
     * given:
     *
     *	ptr	    - pointer to start of json blob
     *	len	    - length of the json blob
     *	is_valid    - if != NULL set to true or false depending on json validity
     *
     * Returns a pointer to the parsed json tree.
     *
     * XXX - should this function return on error ? - XXX
     */
    struct json *
    json_parse(char const *ptr, size_t len, bool *is_valid)
    {
    }

There's one sanity check I'm not sure on which I've added an XXX comment
to.