JSON parser no longer rejects embedded NUL bytes
That is to say that the is_string() function is no longer called on the
string to parse. This does NOT mean that parsing of NUL bytes is done:
it just means that a string with NUL bytes in it will not be rejected
outright. More has to be done here.