Change debug output of parse_json_(block|file)
Instead of (in jparse.y main()) saying that it's about to call the
parse_json_file() the parse_json_file() now says it's about to call
parse_json_block() so that it will be more clear where the code is. Also
the debug output now properly shows the length in the call which was
added in commit c560d71c (this is done for when directly calling
parse_json_block() as well via the -s option to jparse).