JSON parser now allows NUL bytes - resolve #208
If a string (or a file) has a NUL byte not within a string it is a
syntax error in the parser and parsing will not continue. If however a
NUL byte is in a string it will be accepted.

The way this works is quite simple. The string regex already allowed for
NUL bytes and the default rule (that we define later) in the lexer does
not allow for NUL bytes. Now in the lexer we have the length of each
pattern that matched in ugly_leng. This variable does not exist in bison
though so for each string matched we set ugly_length to ugly_leng
(perhaps this should always be done?). Now the parse_json_string()
function has changed to be:

    /* parse_json_string - parse a json string
     *
     * given:
     *
     *	string	    - the text that triggered the action
     *	len	    - length of the string to convert (important for NUL bytes)
     *
     * Returns a pointer to a struct json with the converted string.
     *
     * NOTE: The len is important for strings that have NUL bytes: without it we
     * would rely on strlen() which would mean that the first NUL byte would be the
     * end of the string. If len <= 0 this function uses strlen() on the string.
     *
     * NOTE: This function does not return if passed a NULL string or if conversion
     * fails.
     */
    struct json *
    parse_json_string(char const *string, size_t len)
    {
    }

Then in the bison grammar rule for string the action looks like:

    json_string: JSON_STRING
		    {
			$$ = parse_json_string(ugly_text, ugly_length);
			json_dbg(JSON_DBG_MED, __func__, "under json_string: after JSON_STRING returning type: %s",
					   json_element_type_name($$));
			json_dbg(JSON_DBG_LOW, __func__, "under json_string: after JSON_STRING before ;");
		    }
		    ;

The important line is the one that calls parse_json_string():

    $$ = parse_json_string(ugly_text, ugly_length);

An important thing to note is that when the debug functions show what is
about to be parsed has a NUL byte it will stop at the first NUL so that
for a file that has a NUL byte between two "s you would see:

    JSON DEBUG[1]: Calling parse_json_block(""", 4):
    JSON DEBUG[1]: *** BEGIN PARSE:
    <
    "
    >

because the NUL byte after the first '"' makes the debug function think
it's reached the end of the string. There are a few ways to go about
solving this which I'm going to bring up on the GitHub issue as this is
already much too long.

In any case this resolves issue #208 for both files with NUL bytes as
well as strings on the command line with NUL bytes.