Update man pages for FILES description consistency
For the source file of each tool the description is 'Source code of the
foo tool'. For the header file for each tool the description says 'Header
file for the foo'. There's an exception: if a header file is needed for
the tool but not directly it is not worded this way. For example the
limit_ioccc.h file is worded like:

    Header file with limits related to the IOCCC.

I've not updated the dates on these man pages as these changes are only
cosmetic.