New utility functions to print unprintable bytes
In the case of C \-escape characters it prints them as a literal string
(except that \n and \t are actually shown as a newline and a tab). For
!isprint() it shows them as the bytes themselves. Anything else is shown
as they are normally. For example [0-9a-zA-Z] are shown exactly as those
characters. The functions are:

    /* fprint_str
     *
     * Print string str to a FILE * with indistinct chars represented in a way that
     * can be seen
     *
     * given:
     *      stream  - FILE * to write to
     *      str     - the string to print
     *      len     - the length of the string
     *
     * NOTE: If len == USE_STRLEN (defined in util.h) then strlen() is used to
     * determine the length. If >= 0 it is the caller's responsibility to be
     * absolutely sure that the length does not go beyond the actual string length!
     *
     * This function prints regular characters that are easy to distinguish as
     * normal text (for example characters in the range [0-9a-zA-Z] will be shown as
     * those characters exactly though those aren't the only ones) as they are but
     * where a character is not easy to distinguish (e.g. a NUL byte) it shows the
     * literal character (i.e. "\\0" for "\0") so that it can be seen what the
     * character actually is.
     *
     * NOTE: As for the printf escape chars '\n' is printed as a newline and '\t' is
     * shown as a tab but the others (unless I missed some) are shown as C \-escape
     * strings. Some other characters that are not printable are also shown but as
     * literal bytes.
     *
     * NOTE: This function will return false on a NULL string but will not return on
     * NULL stream.
     *
     * NOTE: This function returns true if everything was successfully printed but
     * it might be nice if it took another arg, a bool *, which would say if any
     * unprintable characters were printed.
     *
     * XXX A version that returns a char * might be useful in some cases ?
     */
    bool
    fprint_str(FILE *stream, char const *str, ssize_t len)
    {
    }

    /* print_str    - wrapper to fprint_str() that writes to stdout
     *
     * given:
     *
     *      str     - the string to print
     *      len     - the length of the string (-1 == use strlen(), >= 0 is the
     *                exact length)
     *
     * NOTE: This just returns fprint_str(). We let that function do all the checks.
     */
    bool
    print_str(char const *str, ssize_t len)
    {
    }

As noted in the comments the function returns true if there were no
errors but perhaps it could return instead if it printed any bytes that
are !isprint(). Alternatively a bool * could be passed in to indicate
this.

The function parse_json_block() uses this facility. Note though that
parse_json_string does NOT which means that a string with a NUL byte
will still look confusing. This is an example of where a version that
returns a char * would be useful but that's more complicated and perhaps
simply restructuring the calls to the json_dbg() function is all that's
needed but this is to be decided later on.

If a file has (in C form):

    "a NUL \0 byte"

then that is what the above functions will print.