Move JSON conversion functions to json_util.{c,h}
The tree walk functions along with the json free functions have been
moved as well. As these are utility functions for the json parser it
seems the more appropriate location. I thought of moving some
additional functions over like json_decode() but this would make json.c
almost empty if not empty and since the functions moved are for the
parser (whereas the others are not) it seems appropriate to have them in
json_util but not the others.

Makefile dependencies update.