Fix parsing JSON members + improve debug output
Even debug statements cannot be in mid-rule actions in bison so I've
reordered the debug statements a bit with some improvements as well as
below. This fixes it so that JSON members can be parsed.

Show $$ and $1 (and in one case $3) types at appropriate place.

Show function calls in debug statements with (for now) the type as the
parameter (exceptions: numbers and strings are the number string and the
string and length).

There might be some inconsistencies to get rid of and there is
definitely room for improvement. For example showing the value of the
types would be ideal. For some (e.g. in parse_json_string()) the
conversion functions could show it (like it does for type in
parse_json_member()).

Improve error messages in parse_json_member (for NULL pointers). This
allows for knowing which pointer is NULL (this helped identify the
problem being fixed in this commit).

Sequenced exit codes.