Rename dashed files to underscored files
For the most part, we use filenames with _ (underscore)
instead of - (dash).

It can be argued which character is better (we tend to prefer
- (dash) because it doesn't require the use of a shift key),
however having a mix is sub-optimal and forces one to guess
which character is in the filename.

Given that more filenames used _ (underscore) instead of
- (dash), we chose to make the more minimal change.

This "yak shave" came about because we guessed wrong
one too may times.  :-)