writeable -> writable
Contrary to what vim thinks the word is 'writable' is not a regional
spelling for 'writeable'. The word is 'writable' full stop. That
/usr/share/dict/words has it in under linux and macOS does not change
the fact it's both incorrect and not even a word so spelt (iOS also has
both, interestingly) even if it might make more sense (to some people).
The thing is spelling often does not make sense. For an amusing and
ironic example: short is longer than long. Possibly it's modelled after
other words like uniteable (alternative spelling for unitable though vim
does not recognise either word and neither do some other dictionaries
but both mean some things/people that can be united). But I've digressed
well beyond the point due to my love of words. The point is writable is
the correct spelling.