Remove redundant debug statements in jparse.l
The exceptions being whitespace (which will eventually be removed from
debug output and maybe could be already) and an invalid token (as the
error in jparse.y does not include that - for now at least). This helps
keep the output of jparse_test.sh more consistent (it showed the debug
output and then after that it would show the debug statements in
jparse.l).

I also removed the variable 'int token' as it's not used and I don't
think it will be needed after all (I always doubted or at least
questioned whether it will be needed).