Bison rules now use named references
Instead of the positional references we use named references. This makes
the intent clearer and it also makes it easier to verify that we have
things right.

There are currently two cases where we cannot use them and have to use
positional references. This is due to ambiguity.

TODO:

- Verify that the numbers are all correct (see below).
- Remove any unnecessary debug statements with these references (in some
  cases they're not needed as they're not actually there i.e. they're
  NULL).
- Improve debug messages to help understand the actions better.

As for verifying the numbers are all correct observe the following rule:

    json_member:        json_string
			    JSON_COLON
			json_element
			    {
				/*
				 * $$ = $json_member
				 * $1 = $json_string
				 * $3 = $json_element
				 */
				json_dbg(JSON_DBG_MED, __func__, "under json_member: json_string JSON_COLON json_element:");
				json_dbg(JSON_DBG_LOW, __func__, "$json_member = parse_json_member($json_string = %s, $json_element = %s)",
				    json_element_type_name($json_string), json_element_type_name($json_element));
				$json_member = parse_json_member($json_string, $json_element, &tree);
				json_dbg(JSON_DBG_LOW, __func__, "under json_member: $json_member type: %s", json_element_type_name($json_member));
				json_dbg(JSON_DBG_MED, __func__, "under json_member: after json_element returning type: %s",
						       json_element_type_name($json_member));
				json_dbg(JSON_DBG_LOW, __func__, "under json_member: after json_element before ;");
			    }
			;

The comments I added indicate what the positional references actually
are but there might be cases where this is incorrect (in at least one
case I had used $2 when it was actually $3 but there might be others
like this). These comments are to help follow it better as well as
verify that the references are correct but eventually they can be
removed.