Grammar fixes
Where possible surround dots with 's. This is not always practical or
even possible (e.g. commands) but I have changed jnum_gen.c to have:

    -		   "\t/* true ==> as_str had a . in it such as 1.234, false ==> no . found */\n",
    +		   "\t/* true ==> as_str had a '.' in it such as 1.234, false ==> no '.' found */\n",

so that the vim spellcap is not triggered and also to make it clearer
what it being said. I also did make rebuild_jnum_test to update
jnum_test.c.

There are almost certainly other places where this can still be done
(and this will be necessary in the future as well) and not just dots or
full stops: commas in some cases too e.g. when expecting a comma in some
text.

I also removed extraneous (which were also erroneous) '...' from the
usage message of both jinfochk and jauthchk (in the header files and the
man pages).