Bug fix and enhance warn and debug functions
In some functions the check if output was allowed was done prior to
checking if that bool should be set to false. I believe this was done
when I moved the check up to make it not do anything at all if it was
not allowed.

With the exception of warn_or_err functions they all now return a bool
indicating whether the output was allowed or not. This will be important
for a future commit (with json debugging).

I removed extraneous function prototypes of the json jwarn() and jerr()
functions from json_chk.h (the functions are in json_util.c so now
they're only in json_util.h).