JSON parser now takes a struct json *
The parser is now passed a struct json *node that can be accessed in the
parser actions. Currently the struct json *node from within parse_json()
is NULL but this can be changed so that the parser can create the parse
tree. The way this is done is by specifying in jparse.y at the
declarations:

    /*
     * we need access to the node in parse_json() so we tell bison that ugly_parse()
     * takes a struct json *node.
     */
    %parse-param { struct json *node }

This changes the functions ugly_parse() and ugly_error() to be:

    int ugly_parse (struct json *node);

and

    void ugly_error(struct json *node, char const *format, ...);

In the lexer where we call ugly_error() (on invalid token - which
probably could be and will be removed in time) we pass NULL as the first
argument and we don't act on it there since it's NULL (it checks for
NULL though with the function it uses it would be safe to call it
anyway).

In parse_json() in jparse.l we call ugly_parse() like:

    /*
     * parse the blob, passing into the parser the node. The parser can access
     * node directly by referring to node. For example to print out the type
     * (currently it'll only be JTYPE_UNSET as it's NULL) it could do:
     *
     *  json_dbg(JSON_DBG_LOW, __func__, "node type: %s", json_element_type_name(node));
     *
     */
    ret = ugly_parse(node);

and in the parser (as described in the comment) we can directly refer to
node to access the root of the tree.

Removed the 'tree' parameter from the parse_json_() functions.

This feels worthy of updating the jparse version so it's now version 0.2
2022-05-22.

Note that when making the parser re-entrant some of this will have to
change (as the parser function will take additional parameters) but we
can worry about that another time.