Reverse commit on the debug facility
While the idea of returning "allowed boolean" is interesting,
the impact on the dbg facility is incompatible with the
the intent of the debug facility in which simple calls
to debugging functions can me made.

By not returning void, these calls would suggest that the caller,
in order to take advantage of the "allowed boolean", the caller
would have to either assign the return value to a boolean (and thus
have to change the caller's stack) which in could modify the caller
function environment.  Such a stack modification could change a
bug's behaviour making it harder to debug (creating the potential
for an "Heisenbug").

Even if the caller were to directly test the "allowed boolean"
within some if statement, the caller need to do something with
the result of the if statement, changing behaviour of the calling
function.

The dbg facility is widely used in other tools.  We would prefer
to not fork the dbg facility (creating two different behaviours)
in this repo for such a purpose.

Therefore with regret, we reverse commit cab3cd304ff786be747827dae426fd750612a1b4
for dbg.c and dbg.h.